fix: ssl and backup translations
