+++ packages/manager/modules/pci/src/components/project/storages/databases/database.class.js
@@ -1,11 +1,6 @@
 import find from 'lodash/find';
 
 import Base from './base.class';
-import {
-  SSL_MODE_REQUIRED,
-  SSL_MODE_NA,
-  SSL_MODE_SSL_TLS,
-} from './databases.constants';
 
 import { ENGINES_NAMES } from './engines.constants';
 
@@ -105,22 +100,6 @@ export default class Database extends Base {
       );
     }
     return this.currentEngine;
-  }
-
-  getSSLModeKey() {
-    if (!this.sslModeKey) {
-      this.sslModeKey = this.sslMode;
-      if (SSL_MODE_REQUIRED.includes(this.sslMode)) {
-        this.sslModeKey = 'required';
-      }
-      if (SSL_MODE_NA.includes(this.sslMode)) {
-        this.sslModeKey = 'n/a';
-      }
-      if (SSL_MODE_SSL_TLS.includes(this.sslMode)) {
-        this.sslModeKey = 'SSL_TLS';
-      }
-    }
-    return this.sslModeKey;
   }
 
   updateData(data) {
