+++ packages/manager/modules/pci/src/components/project/storages/databases/databases.constants.js
@@ -8,15 +8,6 @@ export const STATUS = {
   UPDATING: 'UPDATING',
 };
 
-export const SSL_MODE_REQUIRED = ['require', 'required', 'REQUIRED'];
-
-export const SSL_MODE_NA = ['n/a'];
-
-export const SSL_MODE_SSL_TLS = ['TLS/SSL'];
-
 export default {
   STATUS,
-  SSL_MODE_REQUIRED,
-  SSL_MODE_NA,
-  SSL_MODE_SSL_TLS,
 };
