+++ packages/manager/modules/pci/src/projects/project/storages/databases/database/general-information/general-information.html
@@ -304,7 +304,7 @@
                 >
                     <oui-tile-description>
                         <span
-                            data-translate="{{'pci_databases_general_information_' + $ctrl.database.getSSLModeKey($ctrl.endpoint.sslMode)}}"
+                            data-translate="{{$ctrl.getSSLKeyTranslation()}}"
                         ></span>
                     </oui-tile-description>
                 </oui-tile-definition>
